Feat/v2 (#7)
* fix: no longer using signal

* feat: pthipth init

* fix: function wrong return type in get stack size

* fix: prevent pthipth_init

* demo: compare_create

* fix: add pthipth_init in demo

* chore: move atomic_load outside loop

* fix: change default priority of idle thread

* fix: pthipth_cond

* fix: mutex_count in pthipth_private_t

* fix: not use blocked state in mutex, boost priority to 0 insteads of priority inheritance, and modified try lock

* fix: set detach in first line

* fix: pthipth_init watch the tid

* fix: chekc is deatch when join

* fix: use idle priority size in hash table and prevent prio_delete when delete not exist in bucket

* fix: full queue size yield and does not destroy if have task in queue

+++ src/pthipth_get.c
@@ -1,5 +1,4 @@
 #include "pthipth_internal.h"
-#include "pthipth_signal.h"
 
 extern __thread pthipth_private_t *cur_pthipth;
 
@@ -10,7 +9,7 @@ int pthipth_get_prio()
     return cur_pthipth ? cur_pthipth->priority : - 1;
 }
 
-uint64_t pthipth_get_stack_size()
+int pthipth_get_stack_size()
 {
     return cur_pthipth ? cur_pthipth->stack_size : 0;
 }
